Updated Dimentions
The pin header was up to high. So I fixed that so now it aligns properly with the main computer. I also Fixed the traces so none of the via's sir between any pads, this will decrease the risk of any shorts.

I used the paper printout method to verify my dimensions.